% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.helper.R
\name{connect.helper}
\alias{connect.helper}
\title{connect.helper}
\usage{
connect.helper(nhab, ndest, trans, hpoint, dpoint, maxdist, a)
}
\arguments{
\item{nhab}{Number of habitat cells}

\item{ndest}{Number of destination cells}

\item{trans}{Transition raster}

\item{hpoint}{Habitat points}

\item{dpoint}{Destination points}

\item{maxdist}{Maximum kernel distance}

\item{a}{Parameter -a- for dispersal kernel}
}
\value{
A spatial raster of connectivity.
}
\description{
Generates a connectivity raster based on species habitat locations.
    The connectivity model determines connectivity based on a dispersal kernel,
    obstacles to movement (e.g. land), and habitat quality.
}
