% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral.R
\docType{data}
\name{coral}
\alias{coral}
\title{Raster layer of coral habitats around Zanzibar}
\format{
An object of class RasterLayer
}
\usage{
data(coral)
}
\description{
A small example raster layer of species habitats/distributions
that you could base connectivity models on.
}
\examples{

data(coral)
head(coral)

}
\references{
Sacre et al. (2023)
}
\keyword{datasets}
