% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wander_data.R
\name{wander_data}
\alias{wander_data}
\title{wander_data}
\usage{
wander_data(name)
}
\arguments{
\item{name}{Name of the dataset to be fetched. Options are: "coral" and "zanzibar".}
}
\value{
A raster.
}
\description{
Fetch example data for the wandeR package
}
