% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wander_kernel.R
\name{wander_kernel}
\alias{wander_kernel}
\title{wander_kernel}
\usage{
wander_kernel(
  maxdist = 10000,
  kernel = c("neg_exp", "beta"),
  t = NULL,
  a = NULL,
  b = NULL
)
}
\arguments{
\item{maxdist}{The maximum dispersal distance of the species. Default is
10,000.}

\item{kernel}{The dispersal kernel distribution. Options are 'neg_exp' or
'beta'. Default is 'neg_exp'.}

\item{t}{Parameter for the negative exponential kernel. Values between 0
and 1 give a normal exponential decay kernel, with values closer to 0
giving a steeper decline. Values above 1 give an essentially linear
distribution. Default value is 0.2.}

\item{a}{Parameter for the beta distribution kernel. Default value is 2.}

\item{b}{Parameter for the beta distribution kernel. Default value is 7.}
}
\value{
A plot of specified dispersal kernels.
}
\description{
Generates plots of different dispersal kernels that
    can be implemented in the 'connect' function. There are two dispersal+
    kernel options, a negative exponential (default) or a beta distribution.
    Test different parameter values to inform what values are chosen for the
    connectivity modelling.
}
