% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zanzibar.R
\docType{data}
\name{zanzibar}
\alias{zanzibar}
\title{Raster layer of waters surrounding Zanzibar}
\format{
An object of class RasterLayer
}
\usage{
data(zanzibar)
}
\description{
A small example raster layer of an ocean areas with land barriers
that you could base connectivity models on.
}
\examples{

data(zanzibar)
head(zanzibar)

}
\references{
Sacre et al. (2023)
}
\keyword{datasets}
