% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect()
}
\arguments{
\item{habitats}{A raster of species habitats or known locations.}

\item{destinations}{A raster of the land or seascape. Values of 0 or NA are considered "barriers".
Values greater than 0 are considered valid location to travel through.}
}
\value{
A spatial raster of connectivity.
}
\description{
Generates a connectivity raster based on species habitat locations.
    The connectivity model determines connectivity based on a dispersal kernel,
    obstacles to movement (e.g. land), and habitat quality.
}
