% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(habitats, destinations, maxdist, nthreads = 1)
}
\arguments{
\item{habitats}{A raster of species habitats or known locations.}

\item{destinations}{A raster of the land or seascape. Values of 0 or NA are considered "barriers".
Values greater than 0 are considered valid locations to travel through.}

\item{maxdist}{The maximum dispersal distance of the species in map units.}

\item{nthreads}{Specify the number of threads to use.}
}
\value{
A spatial raster of connectivity.
}
\description{
Generates a connectivity raster based on species habitat locations.
    The connectivity model determines connectivity based on a dispersal kernel,
    obstacles to movement (e.g. land), and habitat quality.
}
