% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(habitats, destinations, maxdist, t = 0.2, nthreads = 1)
}
\arguments{
\item{habitats}{A raster of species habitats or known locations.}

\item{destinations}{A raster of the land or seascape. Values of 0 or NA are considered "barriers".
Values greater than 0 are considered valid locations to travel through.}

\item{maxdist}{The maximum dispersal distance of the species in map units.}

\item{t}{Parameter for the negative exponential kernel. Values between 0 and 1 give a normal exponential decay kernel,
with values closer to 0 giving a steeper decline. Values above 1 give an essentially linear distribution. Default value is 0.2.}

\item{nthreads}{Specify the number of threads to use.}
}
\value{
A spatial raster of connectivity.
}
\description{
Generates a connectivity raster based on species habitat locations.
    The connectivity model determines connectivity based on a dispersal kernel,
    obstacles to movement (e.g. land), and habitat quality.
}
